requirements.txt
Django~=5.0.11
djangorestframework~=3.15.2
djangorestframework_simplejwt~=5.4.0
django-cron~=0.6.0